Dtsam 531 iac system user case allocator org role (#2339)
* DTSAM-531 case-allocator role config for IAC System user

* DTSAM-531 update rule

* DTSAM-531 move IAC system role tests to IacSystemRoleTest.java

---------
